test(server): run the extension-stampede single-flight test under turmoil DST
The extension-stampede test reproduced the burst with 50 OS-thread-concurrent
clients and a real 50ms persist delay so stampeders pile up inside
extend_window before the first persist completes — a timing-dependent setup.

Convert it to run under turmoil: 50 logically concurrent get_ts requests driven
by join_all on the single-threaded deterministic scheduler. The extension_lock
contention (and so the recheck-after-acquire coalescing) is exercised
identically, but deterministically and instantly. The server pushes the clock
past committed_high_water before serve() binds the listener, so the client
cannot issue any request until the window is already exhausted; a connect-failed
batch contacts the server zero times, so retrying the burst until it lands keeps
the "exactly one persist" assertion exact.

Reuses the shared dst harness (serve / client / into_sim_parts) from
test_support. Replaces the real-time original with the deterministic version
(extension_stampede_dst.rs, required-features = ["dst"]).
